Added code to remove broken jobs.

+++ worker.py
@@ -105,18 +105,26 @@ class Worker(Actor):
             self.status = 'posting results'
 
             # send back result
-            with open(output_file_name, 'r') as result_file:
-                self.mgr.broadcast_message(ResultMessage(msg={
+            if not process.returncode:
+                with open(output_file_name, 'r') as result_file:
+                    self.mgr.broadcast_message(ResultMessage(msg={
+                        'job_id': sim['job_id'],
+                        'sim_id': sim['sim_id'],
+                        'user_id':sim['user_id'],
+                        'result':result_file.read(),
+                        'sample':sim['sample']
+                    }))
+
+                os.remove(output_file_name)
+            else:
+                self.mgr.broadcast_message(JobErrorMessage(msg={
                     'job_id': sim['job_id'],
                     'sim_id': sim['sim_id'],
-                    'user_id':sim['user_id'],
-                    'result':result_file.read(),
-                    'sample':sim['sample']
+                    'user_id':sim['user_id']
                 }))
 
             os.remove(xml_name)
             os.remove(jar_name)
-            os.remove(output_file_name)
         except Exception, e:
             print 'Worker error: ', e
             print 'Stack trace:'
